requirements.txt
flask
werkzeug
pillow
pytesseract